Domain and Repos "fixed". Added Commands. CLoses #3.

+++ src/main/java/slaxation/KampfLand/repositories/ObchodRepository.java
@@ -3,5 +3,5 @@ package slaxation.KampfLand.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import slaxation.KampfLand.domain.Obchod;
 
-public interface ObchodRepository extends JpaRepository<Obchod, Long> {
+public interface ObchodRepository extends JpaRepository<Obchod, Integer> {
 }
