+++ src/main/java/slaxation/KampfLand/repositories/PrevadzkaRepository.java
@@ -3,5 +3,5 @@ package slaxation.KampfLand.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import slaxation.KampfLand.domain.Prevadzka;
 
-public interface PrevadzkaRepository extends JpaRepository<Prevadzka, Long> {
+public interface PrevadzkaRepository extends JpaRepository<Prevadzka, Integer> {
 }
